Restored package structure.
So I understand that the current build file messes up when trying to
build from multiple folders. The default java behavior should be each
folder is a package. So IDEs like Eclipse and netbeans can't acount for
all the files being in the same package but in different folders. This
would be easier fixed by reconfiguring the ant buid file rather then
grouping everything in one package. The alternative is what we did on
last project where we maintained 2 codebases but that is going to get
harder and harder to do as the assignments get trickier as we will have
to handle migrations to the code manually.